pyproject.toml
[project]
name = "llm-pocs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.49.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "pylance>=0.39.0",
    "streamlit>=1.51.0",
]
